First working version
First working version and the tool itself can return cached credentials
within at most 347ms on my machine:

Benchmark 1: 10
  Time (mean ± σ):     288.9 ms ±  26.3 ms    [User: 123.4 ms, System: 17.1 ms]
  Range (min … max):   251.5 ms … 347.5 ms    10 runs

However despite this when ran by kubeconfig's exec we get:

Benchmark 1: 20
  Time (mean ± σ):      2.479 s ±  0.085 s    [User: 0.951 s, System: 0.301 s]
  Range (min … max):    2.344 s …  2.627 s    10 runs

Yet if I output the JSON to a file and use cat to read the JSON for
kubeconfig's exec we get:

Benchmark 1: 20
  Time (mean ± σ):     891.8 ms ±  50.1 ms    [User: 106.1 ms, System: 213.9 ms]
  Range (min … max):   829.4 ms … 956.1 ms    10 runs

So there appears to be an issue of performance when writing to
kubeconfig. I have tried switching from println! to printing to writing
directly to stdout but this has made no difference.